+++ Project.toml
@@ -2,7 +2,7 @@
 name = "DashArcgisOpen"
 uuid = "1b08a953-4be3-4667-9a23-bc02c41c7dc4"
 authors = ["James Holcombe  <jholcombe@hotmail.co.uk>"]
-version = "0.0.2"
+version = "0.0.3"
 
 [deps]
 Dash = "1b08a953-4be3-4667-9a23-3db579824955"
