+++ R/internal.R
@@ -1,12 +1,12 @@
 .dashArcgisOpen_js_metadata <- function() {
 deps_metadata <- list(`dash_arcgis_open` = structure(list(name = "dash_arcgis_open",
-version = "0.0.2", src = list(href = NULL,
+version = "0.0.3", src = list(href = NULL,
 file = "deps"), meta = NULL,
 script = 'dash_arcgis_open.min.js',
 stylesheet = NULL, head = NULL, attachment = NULL, package = "dashArcgisOpen",
 all_files = FALSE), class = "html_dependency"),
 `dash_arcgis_open` = structure(list(name = "dash_arcgis_open",
-version = "0.0.2", src = list(href = NULL,
+version = "0.0.3", src = list(href = NULL,
 file = "deps"), meta = NULL,
 script = 'dash_arcgis_open.min.js.map',
 stylesheet = NULL, head = NULL, attachment = NULL, package = "dashArcgisOpen",
