+++ src/DashArcgisOpen.jl
@@ -3,11 +3,11 @@ module DashArcgisOpen
 using Dash
 
 const resources_path = realpath(joinpath( @__DIR__, "..", "deps"))
-const version = "0.0.2"
+const version = "0.0.3"
 
-include("jl/basemapgallery.jl")
 include("jl/geojsonlayer.jl")
 include("jl/mapview.jl")
+include("jl/basemapgallery.jl")
 
 function __init__()
     DashBase.register_package(
@@ -18,14 +18,14 @@ function __init__()
             [
                 DashBase.Resource(
     relative_package_path = "dash_arcgis_open.min.js",
-    external_url = "https://unpkg.com/dash_arcgis_open@0.0.2/dash_arcgis_open/dash_arcgis_open.min.js",
+    external_url = "https://unpkg.com/dash_arcgis_open@0.0.3/dash_arcgis_open/dash_arcgis_open.min.js",
     dynamic = nothing,
     async = nothing,
     type = :js
 ),
 DashBase.Resource(
     relative_package_path = "dash_arcgis_open.min.js.map",
-    external_url = "https://unpkg.com/dash_arcgis_open@0.0.2/dash_arcgis_open/dash_arcgis_open.min.js.map",
+    external_url = "https://unpkg.com/dash_arcgis_open@0.0.3/dash_arcgis_open/dash_arcgis_open.min.js.map",
     dynamic = true,
     async = nothing,
     type = :js
