+++ src/lib/components/layers/FeatureLayer.js
