+++ src/lib/components/layers/GeoJSONLayer.js
