+++ src/lib/components/widgets/BasemapGallery.js
